Fix a bug
Note should be an empty string, because when a module doesn't have a note, it will be `None` instead of `""`. This makes Better Lectio break if the module doesn't have a note. Instead of fixing it on Better Lectio's side, it should just be standardized like the `lektier` entry being `""` too.